Split the ChoirGroupScreen into three screens
